tools_helpers/retriever_archiver/sample/requirements.txt
# Core dependencies
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.0
tiktoken==0.5.2

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.2
PyPDF2==3.0.1

# NLP and embeddings
sentence-transformers==2.3.1
numpy==1.26.3

# Utilities
tqdm==4.66.1
rich==13.7.0
pydantic==2.5.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
